add byte threshold and remove approximate threshold

+++ components/raftstore/src/store/worker/split_controller.rs
@@ -99,8 +99,6 @@ pub struct RegionInfo {
     pub qps: usize,
     pub peer: Peer,
     pub key_ranges: Vec<KeyRange>,
-    pub approximate_size: u64,
-    pub approximate_key: u64,
     pub flow: FlowStatistics,
 }
 
@@ -111,8 +109,6 @@ impl RegionInfo {
             qps: 0,
             key_ranges: Vec::with_capacity(sample_num),
             peer: Peer::default(),
-            approximate_size: 0,
-            approximate_key: 0,
             flow: FlowStatistics::default(),
         }
     }
@@ -196,7 +192,7 @@ impl Recorder {
             samples,
             config.split_balance_score,
             config.split_contained_score,
-            config.sample_threshold,
+            config.sample_threshold as i32,
         )
     }
 
@@ -329,11 +325,6 @@ impl AutoSplitController {
         let capacity = read_stats_vec.len();
         for read_stats in read_stats_vec {
             for (region_id, region_info) in read_stats.region_infos {
-                if region_info.approximate_size < self.cfg.size_threshold
-                    && region_info.approximate_key < self.cfg.key_threshold
-                {
-                    continue;
-                }
                 let region_infos = region_infos_map
                     .entry(region_id)
                     .or_insert_with(|| Vec::with_capacity(capacity));
@@ -341,6 +332,14 @@ impl AutoSplitController {
             }
         }
         region_infos_map
+    }
+
+    fn read_flow(&self, region_infos: &[RegionInfo]) -> usize {
+        let mut flow = 0;
+        for region_info in region_infos {
+            flow += region_info.flow.read_bytes;
+        }
+        flow / self.cfg.detect_times as usize
     }
 
     pub fn flush(&mut self, read_stats_vec: Vec<ReadStats>) -> (Vec<usize>, Vec<SplitInfo>) {
@@ -351,13 +350,12 @@ impl AutoSplitController {
         for (region_id, region_infos) in region_infos_map {
             let pre_sum = prefix_sum(region_infos.iter(), RegionInfo::get_qps);
             let qps = *pre_sum.last().unwrap(); // region_infos is not empty
-            if qps < self.cfg.qps_threshold {
+            if qps < self.cfg.qps_threshold
+                && self.read_flow(&region_infos) < self.cfg.byte_threshold
+            {
                 self.recorders.remove_entry(&region_id);
                 continue;
             }
-
-            let approximate_keys = region_infos[0].approximate_key;
-            let approximate_size = region_infos[0].approximate_size;
 
             let num = self.cfg.detect_times;
             let recorder = self
@@ -386,9 +384,7 @@ impl AutoSplitController {
                     info!(
                         "load base split region";
                         "region_id"=>region_id,
-                        "size"=>approximate_size,
-                        "keys"=>approximate_keys,
-                        "qps"=>qps
+                        "qps"=>qps,
                     );
                 }
                 self.recorders.remove(&region_id);
@@ -537,18 +533,73 @@ mod tests {
             ],
             vec![b"b", key_b.as_encoded()],
         );
+
+        // test distribution with contained key
+        for _i in 0..100 {
+            let key_ranges = vec![
+                build_key_range(b"a", b"k", false),
+                build_key_range(b"b", b"j", false),
+                build_key_range(b"c", b"i", false),
+                build_key_range(b"d", b"h", false),
+                build_key_range(b"e", b"g", false),
+                build_key_range(b"f", b"f", false),
+            ];
+            check_split(
+                b"isosceles triangle",
+                vec![gen_read_stats(1, key_ranges)],
+                vec![],
+            );
+
+            let key_ranges = vec![
+                build_key_range(b"a", b"f", false),
+                build_key_range(b"b", b"g", false),
+                build_key_range(b"c", b"h", false),
+                build_key_range(b"d", b"i", false),
+                build_key_range(b"e", b"j", false),
+                build_key_range(b"f", b"k", false),
+            ];
+            check_split(
+                b"parallelogram",
+                vec![gen_read_stats(1, key_ranges)],
+                vec![],
+            );
+
+            let key_ranges = vec![
+                build_key_range(b"a", b"l", false),
+                build_key_range(b"a", b"m", false),
+            ];
+            check_split(
+                b"right-angle trapezoid",
+                vec![gen_read_stats(1, key_ranges)],
+                vec![],
+            );
+
+            let key_ranges = vec![
+                build_key_range(b"a", b"l", false),
+                build_key_range(b"b", b"l", false),
+            ];
+            check_split(
+                b"right-angle trapezoid",
+                vec![gen_read_stats(1, key_ranges)],
+                vec![],
+            );
+        }
     }
 
     fn check_split(mode: &[u8], qps_stats: Vec<ReadStats>, split_keys: Vec<&[u8]>) {
         let mut hub = AutoSplitController::new(SplitConfigManager::default());
         hub.cfg.qps_threshold = 1;
         hub.cfg.sample_threshold = 0;
-        hub.cfg.key_threshold = 0;
-        hub.cfg.size_threshold = 0;
 
         for i in 0..10 {
             let (_, split_infos) = hub.flush(qps_stats.clone());
             if (i + 1) % hub.cfg.detect_times == 0 {
+                assert_eq!(
+                    split_infos.len(),
+                    split_keys.len(),
+                    "mode: {:?}",
+                    String::from_utf8(Vec::from(mode)).unwrap()
+                );
                 for obtain in &split_infos {
                     let mut equal = false;
                     for expect in &split_keys {
@@ -573,8 +624,6 @@ mod tests {
         hub.cfg.qps_threshold = 2000;
         hub.cfg.sample_num = 2000;
         hub.cfg.sample_threshold = 0;
-        hub.cfg.key_threshold = 0;
-        hub.cfg.size_threshold = 0;
 
         for _ in 0..100 {
             // qps_stats_vec contains 2000 qps and a readStats with a key range;
@@ -590,49 +639,6 @@ mod tests {
             }
             qps_stats_vec.push(qps_stats);
             hub.flush(qps_stats_vec);
-        }
-    }
-
-    fn add_region(read_stats: &mut Vec<ReadStats>, id: u64, key: u64, size: u64) {
-        let mut region_info = RegionInfo::new(read_stats[0].sample_num);
-        region_info.approximate_key = key;
-        region_info.approximate_size = size;
-        read_stats[0].region_infos.entry(id).or_insert(region_info);
-    }
-
-    #[test]
-    fn test_threshold() {
-        let hub = AutoSplitController::new(SplitConfigManager::default());
-        let mut read_stats = vec![ReadStats::default()];
-        add_region(
-            &mut read_stats,
-            1,
-            hub.cfg.key_threshold,
-            hub.cfg.size_threshold,
-        );
-        add_region(
-            &mut read_stats,
-            2,
-            hub.cfg.key_threshold,
-            hub.cfg.size_threshold - 1,
-        );
-        add_region(
-            &mut read_stats,
-            3,
-            hub.cfg.key_threshold - 1,
-            hub.cfg.size_threshold,
-        );
-        add_region(
-            &mut read_stats,
-            4,
-            hub.cfg.key_threshold - 1,
-            hub.cfg.size_threshold - 1,
-        );
-
-        let regions = hub.collect_read_stats(read_stats);
-        assert_eq!(regions.len(), 3);
-        for (id, _) in regions {
-            assert_ne!(id, 4);
         }
     }
